+++ files/urls.py
@@ -9,7 +9,6 @@ urlpatterns = [
     re_path(r"^about", views.about, name="about"),
     re_path(r"^categories$", views.categories, name="categories"),
     re_path(r"^edit", views.edit_media, name="edit_media"),
-    re_path(r"^embed", views.embed_media, name="get_embed"),
     re_path(r"^fu/", include(("uploader.urls", "uploader"), namespace="uploader")),
     re_path(r"^history$", views.history, name="history"),
     re_path(r"^latest$", views.latest_media),